requirements.txt
supabase>=2.8.0
python-dotenv==1.0.0
requests==2.31.0
pandas>=1.5.0
tqdm==4.66.1
wikipedia-api==0.6.0
